requirements.txt
pandas
tqdm
python-dotenv
jupyter
jupyter_http_over_ws
jupyterlab
transformers==4.40.0
datasets
accelerate
scikit-learn
seaborn
matplotlib
tqdm
plotly
google-api-core
faiss-gpu
umap-learn
tabulate
torch
transformers
hydra-core